refactor: dto 디렉토리구조 변경 (request, response 폴더 생성)
